feat: add release-drafter configuration
This commit introduces a new `.github/release-drafter.yml` file which is used to automatically draft new release notes. The template for the release notes is also defined in this file.

feat: add Python package publishing workflow

A new GitHub Actions workflow has been added in `.github/workflows/publish-package.yml`. This workflow is triggered on new tags and is responsible for building and publishing the Python package to PyPI.

feat: add Release Drafter workflow

The `.github/workflows/release-drafter.yml` file has been added in this commit. This workflow updates the draft of the release notes as new pull requests are merged into the master branch.